update: UI 개선 (#27)
